+++ maolan/audio/alsa/in.hpp
@@ -8,7 +8,7 @@ namespace maolan::audio
 class ALSAIn : public ALSA
 {
 public:
-  ALSAIn(const std::string &device, const int &channels,
+  ALSAIn(const std::string &device, const size_t &channels,
          const snd_pcm_uframes_t &frames = 512);
 
   void fetch();
